add a pinotClientInterface, refactor unit test
